requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coloredlogs==14.0
colorgram.py==1.2.0
Cython==0.29.19
d2api==1.1
discord.py==1.3.3
dnspython==1.16.0
dotatools==0.1
emoji==0.5.4
humanfriendly==8.2
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.3.0
motor==2.1.0
multidict==4.7.5
packaging==20.3
Pillow==7.1.2
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
regex==2020.5.14
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
